For correctness, modified implementation of inv_sqrt. Corrected two side light modeling.
git-svn-id: http://foreverbell.0ginr.com/svn/ray_tracer@59 a142d196-329d-4e72-972a-086f8ffe8ca1

+++ ray_tracer/point2D.hpp
@@ -27,7 +27,7 @@ namespace ray_tracer {
 	}
 
 	inline double point2D::inv_length() const {
-		return inv_sqrt((float)(x * x + y * y));
+		return inv_sqrt(x * x + y * y);
 	}
 
 	inline point2D operator*(const point2D &p, double f) {
